Current Version of main multitransformer Model
This model takes the affinity and affinity classification and the protein sequence and smiles strings. It applies a log function to the affinity. The model uses chem bert and protein bert to process the smiles string and protein sequence then the output is combined and fed to a final transformer to predict affinity classification and value.